# api/openapi/shiori.yaml
openapi: 3.1.0
info:
  title: Shiori API
  description: Shiori backend server API for media management and extraction.
  version: 1.0.0
servers:
  - url: /
    description: Current server
paths:
  /health/live:
    get:
      summary: Liveness probe
      operationId: getHealthLive
      tags:
        - Health
      responses:
        '200':
          description: Server is alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: alive
                required:
                  - status

  /health/ready:
    get:
      summary: Readiness probe
      operationId: getHealthReady
      tags:
        - Health
      responses:
        '200':
          description: Server is ready to accept requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready
                required:
                  - status
        '503':
          description: Server is not ready
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/capabilities:
    get:
      summary: Get server capabilities
      operationId: getCapabilities
      tags:
        - System
      responses:
        '200':
          description: Server capabilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitiesResponse'

  /api/v1/media:
    get:
      summary: List library media
      operationId: listMedia
      tags:
        - Library
      responses:
        '200':
          description: A list of media items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
    post:
      summary: Add a new media to the library
      operationId: createMedia
      tags:
        - Library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaCreateRequest'
      responses:
        '201':
          description: Media created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/media/{mediaId}:
    get:
      summary: Get media by ID
      operationId: getMedia
      tags:
        - Library
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The media item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '404':
          description: Media not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/ai/models:
    get:
      summary: List AI models and their loaded status
      operationId: listAIModels
      tags:
        - AI
      responses:
        '200':
          description: A list of AI models configured in the server and their status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIModel'

  /api/v1/ai/models/{modelKey}/load:
    post:
      summary: Load a specific AI model into memory (JIT)
      operationId: loadAIModel
      tags:
        - AI
      parameters:
        - name: modelKey
          in: path
          required: true
          schema:
            type: string
            enum: [tiny, default, quality]
      responses:
        '204':
          description: Model loaded successfully
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '502':
          description: LM Studio unavailable or failed to load
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/jobs/extract:
    post:
      summary: Enqueue an extraction job
      operationId: extractJob
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                target:
                  type: string
                  enum: [media, chapter_list]
              required:
                - url
                - target
      responses:
        '202':
          description: Job enqueued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
        '400':
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  /api/v1/challenges/{token}:
    get:
      summary: Open the temporary human-action controller
      operationId: openHumanChallenge
      tags: [Browser]
      parameters:
        - $ref: '#/components/parameters/ChallengeToken'
      responses:
        '200':
          description: Same-origin HTML controller for the isolated browser session
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            text/html:
              schema: { type: string }
        '404':
          $ref: '#/components/responses/ChallengeNotFound'
        '410':
          $ref: '#/components/responses/ChallengeExpired'
    delete:
      summary: Cancel a pending human-verification session
      operationId: cancelHumanChallenge
      tags: [Browser]
      parameters:
        - $ref: '#/components/parameters/ChallengeToken'
      responses:
        '204':
          description: Challenge cancelled
        '404':
          $ref: '#/components/responses/ChallengeNotFound'
        '410':
          $ref: '#/components/responses/ChallengeExpired'

  /api/v1/challenges/{token}/status:
    get:
      summary: Get the state and expiry of a human-verification session
      operationId: getHumanChallengeStatus
      tags: [Browser]
      parameters:
        - $ref: '#/components/parameters/ChallengeToken'
      responses:
        '200':
          description: Current challenge state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStatus'
        '404':
          $ref: '#/components/responses/ChallengeNotFound'

  /api/v1/challenges/{token}/complete:
    post:
      summary: Ask the backend to verify and resume a human action
      description: Completion succeeds only after the backend confirms that the interactive challenge or login is no longer visible. The extraction worker then reopens the originally requested URL.
      operationId: completeHumanChallenge
      tags: [Browser]
      parameters:
        - $ref: '#/components/parameters/ChallengeToken'
      responses:
        '200':
          description: Challenge verified and extraction resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStatus'
        '409':
          description: The challenge is still visible or is in an incompatible state
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          $ref: '#/components/responses/ChallengeNotFound'
        '410':
          $ref: '#/components/responses/ChallengeExpired'

  /api/v1/challenges/{token}/ws:
    get:
      summary: Upgrade to the same-origin browser screencast WebSocket
      description: Binary messages are JPEG frames; client messages are validated viewport, pointer, wheel, and keyboard input events.
      operationId: streamHumanChallenge
      tags: [Browser]
      parameters:
        - $ref: '#/components/parameters/ChallengeToken'
      responses:
        '101':
          description: WebSocket protocol switched
        '403':
          description: Cross-origin WebSocket request rejected
        '404':
          $ref: '#/components/responses/ChallengeNotFound'
        '409':
          description: Another controller is already connected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  parameters:
    ChallengeToken:
      name: token
      in: path
      required: true
      description: Opaque, short-lived, single-purpose browser capability token
      schema:
        type: string
        format: uuid

  responses:
    ChallengeNotFound:
      description: Challenge token not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ChallengeExpired:
      description: Challenge expired or was cancelled
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
      required:
        - title
        - status

    Capability:
      type: object
      properties:
        name:
          type: string
        available:
          type: boolean
      required:
        - name
        - available

    ChallengeStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, verifying, completed, cancelled, expired]
        kind:
          type: string
          enum: [challenge, login]
          description: Human action that must be completed in the isolated browser session.
        connected:
          type: boolean
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
      required: [status, kind, connected, created_at, expires_at]

    CapabilitiesResponse:
      type: object
      properties:
        profile:
          type: string
          enum: [portable, docker]
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/Capability'
      required:
        - profile
        - capabilities

    Media:
      type: object
      properties:
        id:
          type: string
        source_url:
          type: string
          format: uri
        type:
          type: string
          enum: [manga, manhwa, comic, novel, anime, unknown]
        title:
          type: string
        alternative_titles:
          type: array
          items:
            type: string
        description:
          type: string
        cover_url:
          type: string
        authors:
          type: array
          items:
            type: string
        artists:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ongoing, completed, hiatus, cancelled, unknown]
        genres:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - alternative_titles
        - authors
        - artists
        - status
        - genres
        - created_at
        - updated_at

    MediaCreateRequest:
      type: object
      properties:
        source_url:
          type: string
          format: uri
        type:
          type: string
          enum: [manga, manhwa, comic, novel, anime, unknown]
        title:
          type: string
        alternative_titles:
          type: array
          items:
            type: string
        description:
          type: string
        cover_url:
          type: string
        authors:
          type: array
          items:
            type: string
        artists:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ongoing, completed, hiatus, cancelled, unknown]
        genres:
          type: array
          items:
            type: string
      required:
        - type
        - title

    AIModel:
      type: object
      properties:
        key:
          type: string
          enum: [tiny, default, quality]
          description: The alias key used in Shiori
        name:
          type: string
          description: The actual model name configured
        is_loaded:
          type: boolean
          description: Whether the model is currently loaded in LM Studio memory
      required:
        - key
        - name
        - is_loaded